feat: tokenize focus ring and only show when focus-visible
The focus ring color is now a color token in colors.scss and the full
focus ring style is a token in tokens.scss.

The focus ring now only appears on :focus-visible in supported
browsers, which will avoid showing the focus ring for most elements when
only using the mouse.