fix(track): maintain correct activity states after undo
When an activity was on, and an undo was short enough that it did not
remove the in-progress event, the activity state for that activity after
the undo would incorrectly be off instead of on, even though the last
timepair had -1 as its end time. This change checks for that case and
sets the activity state accordingly.