about/Models: scale "learn more about this model" link correctly
This link was sharing its width with the image, which was capped
to avoid huge images on wide screens. This caused the link to float
in the center of the page on wider screens. By excluding this width
rule for the link, it will always be full width and remain aligned left.