TutorialMedia: fix video autoplay on iPhone
Webkit's rules for video autoplaying seem to be inconsistent
across different versions and devices, but adding `playsinline`
fixed things for me in the simulator, and doesn't seem to break
anything on android.

If the video does fail to autoplay for whatever reason, we're
now catching the error, and providing a (admittedly not particularly
elegant) fallback to start playback by tapping.

Safari/iPhone also doesn't care about the video container's border
radius, so we're setting this explicitly on the video element as well.