DevTools: Avoid calling onDestroy after initialization
For the `showCanvasTimelineOnProjectPage` config checkbox, we were
creating a writable store and syncing it to the config var.
Unfortunately, because the first fetch of this value was async,
we were registering our destruction handler after the component
was done initializing, which would throw an error into the console.
(It's not clear to me whether this actually caused a behavior issue
-at most a tiny memory leak from not cleaning up these stores properly.)

Anyway, that's fixed, by using a button, which is more consistent with
the other variables that can be toggled on this page anyway.