ProjectDetailAsync: Don't write activity time values on input
Calling the same update() function on every input event was causing
Safari to clear input fields where the user had typed a decimal point.
We now avoid setting the value on input, and only set it on blur. But,
we can still update the total values on input, so that we can enable the
save button when the total is correct.

Closes #99.